aws_mongodb_sample_dir/requirements.txt
aws-cdk-lib==2.90.0
constructs>=10.0.0,<11.0.0
awscdk_resources_mongodbatlas
boto3
python-dotenv